fix: jobaccounting package migration
